feat:完成蜂鸣器驱动

+++ EC_lib/bsp/bsp_delay/bsp_delay.c
@@ -23,7 +23,7 @@ void delayInit(void)
 
 }
 
-void delayUs(uint16_t nus)
+void delayUs(uint32_t nus)
 {
 	
     uint32_t ticks = 0;
