+++ EC_lib/bsp/bsp_delay/bsp_delay.h
@@ -15,7 +15,7 @@
 #include "struct_typedef.h"
 
 extern void delayInit(void);
-extern void delayUs(uint16_t nus);
+extern void delayUs(uint32_t nus);
 extern void delayMs(uint16_t nms);
 #endif
